+++ src/components/header/index.ts
@@ -1,0 +1,3 @@
+import Header from './header';
+
+export default Header;
