+++ src/components/toDoList/addItemForm/index.ts
@@ -1,0 +1,3 @@
+import AddItemForm from './addItemForm';
+
+export default AddItemForm;
