+++ src/components/toDoList/toDoItem/index.ts
@@ -1,0 +1,3 @@
+import ToDoItem from './toDoItem';
+
+export default ToDoItem;
