+++ src/components/toDoList/toDoList/index.ts
@@ -1,0 +1,3 @@
+import ToDoList from './toDoList';
+
+export default ToDoList;
