+++ src/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
-import App from 'app';
+import App from './app';
 
 ReactDOM.render(
   <App/>,
