+++ src/pages/mainPage/index.ts
@@ -1,0 +1,3 @@
+import MainPage from './mainPage';
+
+export default MainPage;
